+++ android/src/main/java/com/cloudcare/ft/mobile/sdk/tracker/reactnative/FTRUMModule.kt
@@ -1,8 +1,11 @@
 package com.cloudcare.ft.mobile.sdk.tracker.reactnative
 
-import android.util.Log
 import com.cloudcare.ft.mobile.sdk.tracker.reactnative.utils.ReactNativeUtils
-import com.facebook.react.bridge.*
+import com.facebook.react.bridge.Promise
+import com.facebook.react.bridge.ReactApplicationContext
+import com.facebook.react.bridge.ReactContextBaseJavaModule
+import com.facebook.react.bridge.ReactMethod
+import com.facebook.react.bridge.ReadableMap
 import com.facebook.react.modules.network.OkHttpClientProvider
 import com.facebook.react.modules.network.ReactCookieJarContainer
 import com.ft.sdk.DetectFrequency
@@ -18,6 +21,12 @@ import okhttp3.OkHttpClient
 
 class FTRUMModule(reactContext: ReactApplicationContext) :
   ReactContextBaseJavaModule(reactContext) {
+  companion object {
+    val RN_DEV_INNER_URL_REGEX = arrayOf(
+      Regex("^http://((10|172|192).[0-9]+.[0-9]+.[0-9]+|localhost|127.0.0.1):808[0-9]/logs$"),//expo
+      Regex("^http://localhost:808[0-9]/(hot|symbolicate|message|inspector).*$")//rn
+    )
+  }
 
   init {
     OkHttpClientProvider.setOkHttpClientFactory {
@@ -35,14 +44,20 @@ class FTRUMModule(reactContext: ReactApplicationContext) :
     return "FTReactNativeRUM"
   }
 
+  //是否是 debug 发送网络请求
+  private fun isDevUrl(text: String, regexArray: Array<Regex>): Boolean {
+    return regexArray.any {it.matches(text) }
+  }
+
   @ReactMethod
   fun setConfig(context: ReadableMap, promise: Promise) {
     val map = context.toHashMap()
     val rumAppId = map["androidAppId"] as String
-    val sampleRate = map["sampleRate"] as Float?
+    val sampleRate = map["sampleRate"] as Double?
     val enableNativeUserAction = map["enableNativeUserAction"] as Boolean?
     val enableNativeUserView = map["enableNativeUserView"] as Boolean?
     val enableNativeUserResource = map["enableNativeUserResource"] as Boolean?
+    val enableResourceHostIP = map["enableResourceHostIP"] as Boolean?
     val monitorType = ReactNativeUtils.convertToNativeInt(map["errorMonitorType"])
     val deviceMonitorType = ReactNativeUtils.convertToNativeInt(map["deviceMonitorType"])
     val detectFrequency = ReactNativeUtils.convertToNativeInt(map["detectFrequency"])
@@ -50,7 +65,7 @@ class FTRUMModule(reactContext: ReactApplicationContext) :
 
     val rumConfig = FTRUMConfig().setRumAppId(rumAppId)
     if (sampleRate != null) {
-      rumConfig.samplingRate = sampleRate
+      rumConfig.samplingRate = sampleRate.toFloat()
     }
 
     if (enableNativeUserAction != null) {
@@ -63,6 +78,9 @@ class FTRUMModule(reactContext: ReactApplicationContext) :
 
     if (enableNativeUserResource != null) {
       rumConfig.isEnableTraceUserResource = enableNativeUserResource
+    }
+    if (enableResourceHostIP != null) {
+      rumConfig.isEnableResourceHostIP = enableResourceHostIP
     }
 
     if (monitorType != null) {
@@ -86,6 +104,12 @@ class FTRUMModule(reactContext: ReactApplicationContext) :
 
     globalContext?.forEach {
       rumConfig.addGlobalContext(it.key, it.value.toString())
+    }
+
+    if (BuildConfig.DEBUG) {
+      rumConfig.setResourceUrlHandler { url ->
+        return@setResourceUrlHandler isDevUrl(url, RN_DEV_INNER_URL_REGEX)
+      }
     }
 
     FTSdk.initRUMWithConfig(rumConfig)