+++ example/src/App.tsx
@@ -9,13 +9,15 @@ import Config from 'react-native-config';
 import RUMScreen from './rum';
 import LogScreen from './logging';
 import TraceScreen from './tracing';
+import WebViewScreen from './webView';
+import LocalWebViewScreen from './localWebView';
 import { styles } from './utils';
 import {FTRumReactNavigationTracking} from './FTRumReactNavigationTracking';
 
 function Home() {
   return (
     <Tab.Navigator>
-      <Tab.Screen name="HomeScreen" component={HomeScreen} options={{ title: 'Home' }}/>
+      <Tab.Screen name="HomeScreen" component={HomeScreen} options={{ title: 'react-navigation' }}/>
       <Tab.Screen name="Messages" component={Messages} options={{ title: 'Message' }}/>
       <Tab.Screen name="Mine" component={Mine} options={{ title: 'Mine' }}/>
 
@@ -45,6 +47,12 @@ class HomeScreen extends React.Component<{ navigation: any }> {
         <Button title='网络链路追踪' onPress={() => navigation.navigate('Trace')} />
         <View  style={styles.space}/>
         <Button title='RUM数据采集' onPress={() => navigation.navigate('RUM')} />
+        <View  style={styles.space}/>
+        <Button title='主动数据同步' onPress={() => FTMobileReactNative.flushSyncData()} />
+        <View  style={styles.space}/>
+        <Button title='WebView' onPress={() => navigation.navigate('WebView')} /> 
+        <View  style={styles.space}/>
+        <Button title='LocalWebView' onPress={() => navigation.navigate('LocalWebView')} />
       </View>
     );
   }
@@ -54,7 +62,7 @@ class Messages extends React.Component<{ navigation: any }> {
         let { navigation } = this.props;
     return (
       <View style={{ flex: 1, alignItems: 'center', padding: 20 }}>
-          <Button title='MessagesDetail' onPress={() => navigation.navigate('Deatil')} />
+          <Button title='MessagesDetail' onPress={() => navigation.navigate('Detail')} />
 
       </View>
     );
@@ -91,14 +99,14 @@ const navigationRef: React.RefObject<NavigationContainerRef<ReactNavigation.Root
 
 function App() {
   return (
-    // react-navigation 
+    // react-navigation
     // 开启 RUM View 采集
     // 将 example 中 FTRumReactNavigationTracking.tsx 文件拖入您的工程；
 
     // 方法一：如果有使用 createNativeStackNavigator(); 创建原生导航堆栈，建议采用 <Stack.Navigator screenListeners={FTRumReactNavigationTracking.StackListener} nitialRouteName='Home'>开启采集，
     //        这样可以统计到页面的加载时长
     // 方法二：如果没有使用 createNativeStackNavigator(); 要在 NavigationContainer 组件中进行开启采集，如下
-  
+
     <NavigationContainer ref={navigationRef} onReady={() => {
       // 方法二：
       // FTRumReactNavigationTracking.startTrackingViews(navigationRef.current);
@@ -109,9 +117,11 @@ function App() {
         <Stack.Screen name='Trace' component={TraceScreen} options={{ title: '网络链路追踪' }} />
         <Stack.Screen name='Log' component={LogScreen} options={{ title: '日志输出' }} />
         <Stack.Screen name='RUM' component={RUMScreen} options={({ title: 'RUM 数据采集' })} />
-        <Stack.Screen name='Deatil' component={MessagesDetail}/>
+        <Stack.Screen name='Detail' component={MessagesDetail} options={{ title: 'Detail' }}/>
         <Stack.Screen name="Messages" component={Messages} options={{ title: 'Message' }}/>
         <Stack.Screen name="Mine" component={Mine} options={{ title: 'Mine' }}/>
+        <Stack.Screen name="WebView" component={WebViewScreen} options={{ title: 'WebView' }}/>
+        <Stack.Screen name="LocalWebView" component={LocalWebViewScreen} options={{ title: 'LocalWebView' }}/>
       </Stack.Navigator>
     </NavigationContainer>
   );