+++ example/src/logging.tsx
@@ -4,7 +4,7 @@
   import { styles} from './utils';
 
   class LogScreen extends React.Component{
-    static options(props) {
+    static options() {
       return {
         topBar: {
           title: {
@@ -50,11 +50,17 @@
       }}/>
 
       </View>
+      <View  style={styles.list}>
+      <Button title="Log Status: custom (string status)" onPress={()=>{
+        this.logging("custom log content","custom");
+      }}/>
+
+      </View>
       </ScrollView>
       )
     }
 
-    logging(message:string,status:FTLogStatus){
+    logging(message:string,status:FTLogStatus|string){
       FTReactNativeLog.logging(message,status,{"logging_property":"rn_demo"});
     }
   }