+++ ios/FTReactNativeRUM.m
@@ -12,6 +12,27 @@
 #import <FTMobileSDK/FTResourceMetricsModel.h>
 #import <FTMobileSDK/FTResourceContentModel.h>
 #import <React/RCTConvert.h>
+BOOL filterBlackResource(NSURL *url){
+    static NSMutableArray *internalDevResourceBlacklist;
+    static dispatch_once_t onceToken;
+    dispatch_once(&onceToken, ^{
+        internalDevResourceBlacklist = [NSMutableArray new];
+        NSError *error = nil;
+        NSString *pattern = @"^http://((10|172|192).[0-9]+.[0-9]+.[0-9]+|localhost|127.0.0.1):808[0-9]/logs$";
+        NSRegularExpression * regularExpress = [NSRegularExpression regularExpressionWithPattern:pattern options:NSRegularExpressionCaseInsensitive error:&error];
+        [internalDevResourceBlacklist addObject:regularExpress];
+        NSString *rn = @"^http://localhost:808[0-9]/(hot|symbolicate|message|inspector|status|assets).*$";
+        NSRegularExpression * rnRegularExpress = [NSRegularExpression regularExpressionWithPattern:rn options:NSRegularExpressionCaseInsensitive error:&error];
+        [internalDevResourceBlacklist addObject:rnRegularExpress];
+    });
+    for (NSRegularExpression *regex in internalDevResourceBlacklist) {
+        NSTextCheckingResult *firstMatch =[regex firstMatchInString:url.absoluteString options:0 range:NSMakeRange(0, [url.absoluteString length])];
+        if (firstMatch) {
+            return YES;
+        }
+    }
+    return NO;
+}
 @implementation FTReactNativeRUM
 RCT_EXPORT_MODULE()
 RCT_REMAP_METHOD(setConfig,
@@ -41,9 +62,17 @@ RCT_REMAP_METHOD(setConfig,
     if ([context.allKeys containsObject:@"detectFrequency"]) {
         rumConfig.monitorFrequency =(FTMonitorFrequency)[RCTConvert int:context[@"detectFrequency"]];
     }
+    if ([context.allKeys containsObject:@"enableResourceHostIP"]) {
+        rumConfig.enableResourceHostIP = [RCTConvert BOOL:context[@"enableResourceHostIP"]];
+    }
     if ([context.allKeys containsObject:@"globalContext"]) {
         rumConfig.globalContext = [RCTConvert NSDictionary:context[@"globalContext"]];
     }
+#if DEBUG
+    rumConfig.resourceUrlHandler = ^BOOL(NSURL * _Nonnull url) {
+        return filterBlackResource(url);
+    };
+#endif
     [[FTMobileAgent sharedInstance] startRumWithConfigOptions:rumConfig];
     resolve(nil);
 }
@@ -122,7 +151,7 @@ RCT_REMAP_METHOD(addResource,
     contentModel.responseHeader = [RCTConvert NSDictionary:content[@"responseHeader"]];
     contentModel.responseBody = [RCTConvert NSString:content[@"responseBody"]];
     contentModel.httpStatusCode = [RCTConvert int:content[@"resourceStatus"]];
-    
+
     [[FTExternalDataManager sharedManager] addResourceWithKey:key metrics:metricsModel content:contentModel];
     resolve(nil);
 }