+++ src/ft_tracing.tsx
@@ -25,21 +25,7 @@ import { NativeModules } from 'react-native';
    enableLinkRUMData?:boolean,
    enableNativeAutoTrace?:boolean,
  };
-/**
- * trace 采集数据
- * @param httpMethod 请求方法
- * @param requestHeader 请求头参数
- * @param statusCode 返回状态码
- * @param responseHeader 返回头参数
- * @param errorMessage 错误消息
- */
- export interface FTTraceResource{
-   httpMethod:string,
-   requestHeader:object,
-   statusCode?:number,
-   responseHeader?:object,
-   errorMessage?:string
- };
+
  type FTReactNativeTraceType = {
   /**
    * 配置 trace 开启链路追踪。
@@ -51,8 +37,15 @@ import { NativeModules } from 'react-native';
    * 获取 trace http 请求头数据。
    * @param url 请求地址
    * @returns trace 添加的请求头参数  
+   * @deprecated use getTraceHeaderFields() replace.
    */
    getTraceHeader(key:String, url: String): Promise<object>;
+   /**
+   * 获取 trace http 请求头数据。
+   * @param url 请求地址
+   * @returns trace 添加的请求头参数  
+   */
+   getTraceHeaderFields(url: String,key?:String): Promise<object>;
  };
 
  class FTReactNativeTraceWrapper implements FTReactNativeTraceType {
@@ -68,8 +61,18 @@ import { NativeModules } from 'react-native';
    * @returns a Promise.
    */
    getTraceHeader(key:String, url: String): Promise<object>{
-     return this.trace.getTraceHeader(key,url);
+     return this.trace.getTraceHeaderFields(url,key);
    }
+    /**
+   * 获取 trace http 请求头数据。
+   * @param url 请求地址
+   * @param key 唯一 id
+   * @returns a Promise.
+   */
+   getTraceHeaderFields(url: String,key?:String): Promise<object>{
+    return this.trace.getTraceHeaderFields(url,key);
+   }
+
  }
  export const FTReactNativeTrace:FTReactNativeTraceType = new FTReactNativeTraceWrapper();